Made APU core channels generate samples in the DSP's native format to begin with, avoiding any uneccessary conversions that may also affect sound quality.

+++ src/include/apu.h
@@ -59,7 +59,7 @@ typedef struct apu_chan_s
 
    BOOL enabled;
 
-   INT32 output;
+   REAL output;
 
    REAL phaseacc;
    int freq;
